+++ src/scriptkit/__init__.py
@@ -30,7 +30,7 @@ from .settings import (
 )
 from .times import timestamp
 
-__version__ = "0.5.2"
+__version__ = "0.5.3"
 
 __all__ = [
     "ENV_PREFIX",